fix(suite-native): global passphrase handler
