+++ suite-native/passphrase/src/index.ts
@@ -2,3 +2,4 @@ export * from './passphraseSlice';
 export * from './passphraseThunks';
 export * from './useHandleDuplicatePassphrase';
 export * from './useFeatureRequestsPassphrase';
+export * from './useHandleDeviceRequestsPassphrase';
